+++ R/random_fact.R
@@ -16,6 +16,12 @@
 #' The available categories can be viewed using \code{fetch_categories()}.
 #' (Currently ignored.)
 #'
+#' @param sound Logical indicatin whether or not "Chuck Norris" should be
+#' shouted when generating random facts. Default is \code{FALSE}.
+#'
+#' @param sound_repeat Integer specifying the number of times the sound should
+#' be repeated. Default is 1. (Only matters if \code{sound = TRUE}.)
+#'
 #' @return An object of class \code{"random_fact"}; essentially a character
 #' vector.
 #'
@@ -31,7 +37,8 @@
 #' # Generate multiple random facts
 #' random_fact(3)
 #' }
-random_fact <- function(n = 1L, include = NULL, exclude = NULL) {
+random_fact <- function(n = 1L, include = NULL, exclude = NULL, sound = FALSE,
+                        sound_repeat = 1) {
 
   # Query the ICNDb API
   # url <- modify_url("http://api.icndb.com/jokes/random/", path = n)
@@ -56,6 +63,18 @@ random_fact <- function(n = 1L, include = NULL, exclude = NULL) {
       ),
       call. = FALSE
     )
+  }
+
+  # Play awesome sound
+  if (sound) {
+    for (i in seq_len(sound_repeat)) {
+      if (Sys.info()["sysname"] == "Darwin") {
+        tuneR::play(tuneR::readMP3("inst/sounds/chuck-norris.mp3"),
+                    player = "/usr/bin/afplay")
+      } else {
+        tuneR::play(tuneR::readMP3("inst/sounds/chuck-norris.mp3"))
+      }
+    }
   }
 
   # Return a random_fact object (a list)