+++ NEWS.md
@@ -1,6 +1,8 @@
 # roundhouse 0.0.1.9000
 
-* Fix URL typo in `DESCRIPTION`
+* Fix URL typo in `DESCRIPTION`.
+
+* New `sound` and `sound_repeat` arguments to `random_fact()` for determining whether or not the phrase "Chuck Norris" should be shouted, and how many times, before generating a random fact.
 
 
 # roundhouse 0.0.1